Rename OTHER scenario filters to TYPE
